Add reducer functions to json-loader uses.
The grouped page on the main server suddenly stopped working correctly
more than a year after json-loader was last modified. It appears to be
the cause of the trouble, since the default reducer overwrites earlier
downloads. Who knows how it kept functioning for so long.

Unfortunately, the problem only manifests on the live server. Testing is
therefore hard. Hopefully this change fixes the problem.