Rethink of enabledp
Enabledp was both a check of the enabled flag and of the direction
status. This is a problem, because some flags that don't use direction
are active in nature and need to check for enablement. So, in addition
to approvedp, we need to check three different things (in different
cases) to check if a particular operation should proceed.

Therefore, added directional-p, to see if the directionality of a
directional has been cancelled, separating it out from the enabledp
check. This means that directional flags need to explicitly perform two
checks. Messy but clearer than the alt.